requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
fontawesome==5.10.1.post1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
objects==0.3.1
psycopg2==2.9.5
python-dotenv==0.21.0
requests==2.28.2
SQLAlchemy==1.4.46
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
